lxc: adjusted libvirt-lxc process to add restore mode to it.
When doing lxc migration, we need to restore the lxc container
in dst host from images that we have stored inside a directory, and not
start the container from scratch. For that reason, we should extend
lxcContainerStart into a more generic one, that either starts a container
from scratch or restores it from a snapshot.

Signed-off-by: Katerina Koukiou <k.koukiou@gmail.com>